Refactoring the UsuarioBean class
- Separated the responsibilities between user authentication and user
record
- Created LoginBean bean, to handle the authentication
- Created UsuarioAutenticadoBean, which store the authenticated user on
the application.
